skal vise utfall i historikken i høyremenyen
